pyproject.toml
[project]
name = "mcp-hub"
version = "0.2.0"
description = "Meta-MCP server that manages multiple backend MCP servers through your host agent"
requires-python = ">=3.10"
dependencies = [
    "mcp>=1.6.0",
    "pyyaml>=6.0",
    "aiosqlite>=0.19.0",
    "httpx>=0.27.0",
]

[project.scripts]
mcp-hub = "src.server:main"

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.backends._legacy:_Backend"

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]
